Pie chart and a table(the table will not show net loss when the category of save is negative)
Codes will shows pie chart and a table. If the category of the save is negative, then save in the table will show negative number. In java file, I keep the variables to assign each category has values in codes. You can change or delete those variables to fit the main codes.