Pie chart and a table 2
These codes will  shows a category of net loss. Net loss will not shows in the pie chart. If the value of the save is positive or 0, the category of net loss is invisible in the table. But, If the value of the save is negative, the save in the table will be 0, and the category of net loss will be visible in the table.